+++ ubo_app/constants.py
@@ -107,3 +107,5 @@ CORE_SERVICE_IDS = [
 TEST_INVESTIGATION_MODE = str_to_bool(
     os.environ.get('UBO_TEST_INVESTIGATION_MODE', 'False'),
 )
+
+WAKE_WORD = os.environ.get('UBO_WAKE_WORD', 'hey pod')