+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - refactor(vscode): use data dir as reported by `platformdirs.user_data_path` to store downloaded assets like vscode binary
 - feat(speech-recognition): add `speech-recognition` service with settings menu to download the vosk model and activate/deactivate the listener - closes #284
 - refactor(rgb-ring): add `RgbRingSequenceAction`, add `as_command` to all `RgbRingCommandAction`s
+- feat(speech-recognition): implement wake word detection, the wake word can be set with `UBO_WAKE_WORD` environment variable, the default is `hey pod` - closes #284
 
 ## Version 1.3.0